+++ src/app/ops/(shell)/agreements/page.tsx
@@ -1,0 +1,5 @@
+import { AgreementsIndex } from "./agreements-index";
+
+export default function OpsAgreementsPage() {
+  return <AgreementsIndex />;
+}